Add compatibility with 'upstart' init daemons
- Added compatibility with 'upstart' init daemons with assistance from Tim
Kramer.  This change adds a new config variable 'ENABLE_PSADWATCHD' that
can be used to disable psadwatchd when deployed with upstart since it
has built-in process monitoring and restarting capabilities.  In addition,
a new init script located at init-scripts/upstart/psad has been added that
is compatible with upstart - this script is meant to be copied to the
/etc/init.d/ directory.