Account for older versions of IPTables::ChainMgr that don't export _cmd
Github user itoffshore reported an issue #19 on Alpine Linux where the
following error was produced in 'psad -L' mode:

[+] Listing chains from IPT_AUTO_CHAIN keywords...

Use of uninitialized value in concatenation (.) or string at
/usr/sbin/psad line 6701.
[*]  -t filter -n -L PSAD_BLOCK_INPUT -v does not look like an iptables
command. at /usr/sbin/psad line 6701.

This error was produced because a pre-1.3 version of IPTables::ChainMgr
is installed on the system, and it doesn't export the _cmd hash key
(which is actually exported first by the IPTables::Parse module).