Set ENABLE_PERSISTENCE to "N" by default:
In the psad.conf file, change the ENABLE_PERSISTENCE default to "N" in
order to (by default) limit psad's memory consumption. The trade off is
that really "low and slow" scans may be missed in exchange for a better
operational model. Note the MAX_SCAN_IP_PAIRS variable can also be used
to control memory consumption if ENABLE_PERSISTENCE is enabled.